backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
django-q2>=1.7,<2.0
requests>=2.31,<3.0
Jinja2>=3.1,<4.0
pytest>=8.0,<9.0
pytest-json-report>=1.5,<2.0
pytest-django>=4.8,<5.0
psycopg2-binary>=2.9,<3.0
django-htmx>=1.19,<2.0
